requirements.txt
pymongo
requests
python-dotenv
os